requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
prometheus-client==0.12.0
python-dotenv==0.19.2
requests==2.26.0
urllib3==1.26.7
websocket-client==1.2.1
Werkzeug==2.0.2
